Moves open/closeModal actions down to the index/milestones routes to fix sticky outlets
